requirements.txt
Flask==2.3.2
flask-cors==4.0.0
requests==2.31.0
pandas==2.0.3
python-dotenv==1.0.0
yfinance==0.2.28
beautifulsoup4==4.12.2
lxml==4.9.3